Fix - Per lap calculations not happening
When start_lights == 1, the data structures were being cleared. And then the player list is populated asynchronously.

Sometimes, the session packet would arrive before the player tables are constructed. In this case, the total_laps for fuel/tyre recommenders will not be set

Now, the total laps will always be set. Only when there is a change, the extrapolator object will recompute